+++ NBB.sln
@@ -328,9 +328,9 @@ Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NBB.Http.Effects.Tests", "t
 EndProject
 Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "MultiTenancy", "MultiTenancy", "{6F4DFEED-A5A5-4725-B5A6-40061198AF1B}"
 EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NBB.MultiTenancy.Data.EntityFramework", "src\MultiTenancy\NBB.MultiTenancy.Data.EntityFramework\NBB.MultiTenancy.Data.EntityFramework.csproj", "{F64E5C01-441F-45FA-B532-6011BB0540F9}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NBB.Data.EntityFramework.MultiTenancy", "src\Data\NBB.Data.EntityFramework.MultiTenancy\NBB.Data.EntityFramework.MultiTenancy.csproj", "{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "NBB.MultiTenancy.Abstractions", "NBB.MultiTenancy.Abstractions\NBB.MultiTenancy.Abstractions.csproj", "{C2E2B12E-BB17-4DB3-B455-BF69E4485A11}"
+Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "NBB.MultiTenancy.Abstractions", "src\MultiTenancy\NBB.MultiTenancy.Abstractions\NBB.MultiTenancy.Abstractions.csproj", "{653022CD-6FE5-4C13-B194-013860AAB691}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -752,14 +752,14 @@ Global
 		{4AE166DE-9E9D-482B-84F6-3556AC743DEF}.Debug|Any CPU.Build.0 = Debug|Any CPU
 		{4AE166DE-9E9D-482B-84F6-3556AC743DEF}.Release|Any CPU.ActiveCfg = Release|Any CPU
 		{4AE166DE-9E9D-482B-84F6-3556AC743DEF}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F64E5C01-441F-45FA-B532-6011BB0540F9}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F64E5C01-441F-45FA-B532-6011BB0540F9}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F64E5C01-441F-45FA-B532-6011BB0540F9}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F64E5C01-441F-45FA-B532-6011BB0540F9}.Release|Any CPU.Build.0 = Release|Any CPU
-		{C2E2B12E-BB17-4DB3-B455-BF69E4485A11}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{C2E2B12E-BB17-4DB3-B455-BF69E4485A11}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{C2E2B12E-BB17-4DB3-B455-BF69E4485A11}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{C2E2B12E-BB17-4DB3-B455-BF69E4485A11}.Release|Any CPU.Build.0 = Release|Any CPU
+		{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10}.Release|Any CPU.Build.0 = Release|Any CPU
+		{653022CD-6FE5-4C13-B194-013860AAB691}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{653022CD-6FE5-4C13-B194-013860AAB691}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{653022CD-6FE5-4C13-B194-013860AAB691}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{653022CD-6FE5-4C13-B194-013860AAB691}.Release|Any CPU.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
@@ -910,8 +910,8 @@ Global
 		{B09FE370-C39D-47EB-A555-275467F2F043} = {90E022FB-CA1B-49DD-9BEA-CE7F8E74E8BB}
 		{4AE166DE-9E9D-482B-84F6-3556AC743DEF} = {B09FE370-C39D-47EB-A555-275467F2F043}
 		{6F4DFEED-A5A5-4725-B5A6-40061198AF1B} = {7311E32F-C1B0-41C9-B5F1-DE9EBB6ABB55}
-		{F64E5C01-441F-45FA-B532-6011BB0540F9} = {6F4DFEED-A5A5-4725-B5A6-40061198AF1B}
-		{C2E2B12E-BB17-4DB3-B455-BF69E4485A11} = {6F4DFEED-A5A5-4725-B5A6-40061198AF1B}
+		{F3F8DFC0-FBD0-4F38-9B0E-8A53D4EE4C10} = {F031DBDA-5D3F-4CE6-B133-09CE7E212584}
+		{653022CD-6FE5-4C13-B194-013860AAB691} = {6F4DFEED-A5A5-4725-B5A6-40061198AF1B}
 	EndGlobalSection
 	GlobalSection(ExtensibilityGlobals) = postSolution
 		SolutionGuid = {23A42379-616A-43EF-99BC-803DF151F54E}
