+++ src/MultiTenancy/NBB.MultiTenancy.Abstractions/NBB.MultiTenancy.Abstractions.csproj
