+++ src/MultiTenancy/NBB.MultiTenancy.Abstractions/Tenant.cs
