--- src/MultiTenancy/NBB.MultiTenancy.Data.Abstractions/NBB.MultiTenancy.Data.Abstractions.csproj
@@ -1,11 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\NBB.MultiTenancy.Abstractions\NBB.MultiTenancy.Abstractions.csproj" />
-  </ItemGroup>
-
-</Project>
