--- src/MultiTenancy/NBB.MultiTenancy.Data.EntityFramework/Extensions/DbContextExtensions.cs
@@ -1,10 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Text;
-
-namespace NBB.MultiTenancy.Data.EntityFramework.Extensions
-{
-    public static class DbContextExtensions
-    {
-    }
-}
