--- src/MultiTenancy/NBB.MultiTenancy.Data.EntityFramework/Extensions/ObjectExtensions.cs
@@ -1,12 +1,0 @@
-﻿namespace NBB.MultiTenancy.Data.EntityFramework.Extensions
-{
-    public static class ObjectExtensions
-    {
-        public static bool IsNullOrDefault<T>(this T value)
-        {
-            return ((object)default(T)) == null ?
-                ((object)value) == null :
-                default(T).Equals(value);
-        }
-    }
-}
