--- src/MultiTenancy/NBB.MultiTenancy.Data.EntityFramework/NBB.MultiTenancy.Data.EntityFramework.csproj
@@ -1,21 +1,0 @@
-﻿<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard2.1</TargetFramework>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <PackageReference Include="Microsoft.EntityFrameworkCore" Version="$(MicrosoftEntityFrameworkPackagesVersion)" />
-    <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="$(MicrosoftEntityFrameworkPackagesVersion)" />
-    <PackageReference Include="Microsoft.CSharp" Version="$(MicrosoftCSharpPackageVersion)" />
-    <PackageReference Include="Scrutor" Version="$(ScrutorPackageVersion)" />
-  </ItemGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\..\..\NBB.MultiTenancy.Abstractions\NBB.MultiTenancy.Abstractions.csproj" />
-    <ProjectReference Include="..\..\Core\NBB.Core.Abstractions\NBB.Core.Abstractions.csproj" />
-    <ProjectReference Include="..\..\Data\NBB.Data.Abstractions\NBB.Data.Abstractions.csproj" />
-  </ItemGroup>
-  
-
-</Project>
